requirements.txt
pytest==8.3.5
PyYAML==6.0.2